move TimeElapsed to common
